@@ -1,19 +1,9 @@
 module bitbucket.org/everymind/evmd-golib
 
-go 1.12
+go 1.13
 
 require (
-	bitbucket.org/everymind/gforce v1.3.10
-	github.com/contribsys/faktory v1.2.0-1
-	github.com/contribsys/faktory_worker_go v1.0.0
-	github.com/go-chi/chi v4.0.2+incompatible // indirect
-	github.com/go-chi/jwtauth v4.0.3+incompatible
-	github.com/go-chi/render v1.0.1
-	github.com/gorilla/mux v1.7.3
-	github.com/jmoiron/sqlx v1.2.1-0.20190826204134-d7d95172beb5
-	github.com/lib/pq v1.2.0
-	github.com/pkg/errors v0.8.1
-	github.com/spf13/cast v1.3.0
-	go.uber.org/multierr v1.1.0
-	google.golang.org/appengine v1.6.1 // indirect
+	github.com/jmoiron/sqlx v1.2.0
+	github.com/lib/pq v1.3.0
+	go.uber.org/multierr v1.4.0
 )
