@@ -3,8 +3,9 @@ package dao
 import (
 	"strings"
 
-	"bitbucket.org/everymind/evmd-golib/db"
 	"github.com/jmoiron/sqlx"
+
+	"bitbucket.org/everymind/evmd-golib/db"
 )
 
 func ExecSFEtlData(conn *sqlx.DB, execID int64, tenantID int, objID int64, reprocessAll bool) error {
