@@ -3,8 +3,9 @@ package model
 import (
 	"database/sql"
 
-	m "bitbucket.org/everymind/evmd-golib/modelbase"
 	"github.com/lib/pq"
+
+	m "bitbucket.org/everymind/evmd-golib/modelbase"
 )
 
 type SchemaObject struct {
