@@ -7,12 +7,13 @@ import (
 	"strconv"
 	"time"
 
+	worker "github.com/contribsys/faktory_worker_go"
+	"github.com/gorilla/mux"
+
 	"bitbucket.org/everymind/evmd-golib/db"
 	fn "bitbucket.org/everymind/evmd-golib/faktory/functions"
 	"bitbucket.org/everymind/evmd-golib/faktory/middleware"
 	"bitbucket.org/everymind/evmd-golib/logger"
-	worker "github.com/contribsys/faktory_worker_go"
-	"github.com/gorilla/mux"
 )
 
 type DBVars struct {
